+++ lib/ruy/adapters.rb
@@ -1,6 +1,2 @@
 require_relative 'adapters/file_adapter'
 require_relative 'adapters/sequel_adapter'
-
-module Ruy
-  module Adapters; end
-end
