+++ spec/lib/ruy/conditions/eq_spec.rb
