+++ spec/lib/ruy/conditions/except_spec.rb
