+++ spec/lib/ruy/conditions/less_than_spec.rb
