+++ spec/lib/ruy/outcome_spec.rb
