+++ spec/lib/ruy/rule_set_spec.rb
