+++ spec/lib/ruy/rule_spec.rb
