+++ spec/lib/ruy/rule_storage_spec.rb
